# service/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Инспектор ИИ",
    "version": "0.1.0",
    "description": "Сверка проектной, рабочей и исполнительной документации. Асинхронная pull-модель: загрузка возвращает process_id, клиент опрашивает статус и забирает протокол."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "nullable": true,
                "additionalProperties": true
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/rules/tests/estimate": {
      "get": {
        "summary": "Примерное время пробного прогона по объектам (эксперт)",
        "tags": [
          "rules"
        ],
        "description": "По последней проверке каждого объекта с протоколом: сколько примерно займёт трасса и вариант правила. Оценка — по прошлым прогонам на проверке, а без них — по числу страниц. Ожидание очереди в неё не входит.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "processes"
                  ],
                  "properties": {
                    "processes": {
                      "type": "array",
                      "description": "последняя проверка каждого объекта с протоколом — на них идёт прогон по всем объектам",
                      "items": {
                        "type": "object",
                        "required": [
                          "process_id",
                          "object_id",
                          "pages",
                          "trace_s",
                          "variant_s",
                          "measured"
                        ],
                        "properties": {
                          "process_id": {
                            "type": "string"
                          },
                          "object_id": {
                            "type": "string"
                          },
                          "pages": {
                            "type": "integer",
                            "description": "страниц в принятых файлах проверки"
                          },
                          "trace_s": {
                            "type": "number",
                            "description": "примерное время трассы, секунд"
                          },
                          "variant_s": {
                            "type": "number",
                            "description": "примерное время варианта, секунд"
                          },
                          "measured": {
                            "type": "boolean",
                            "description": "оценка по прошлым прогонам на этой проверке, а не по числу страниц"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rules/tests/{id}": {
      "get": {
        "summary": "Ход и итог пробного прогона правила (эксперт)",
        "tags": [
          "rules"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "code",
                    "variant",
                    "trace",
                    "status",
                    "total",
                    "done",
                    "created_at",
                    "results",
                    "ahead"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "variant": {
                      "type": "object",
                      "nullable": true,
                      "additionalProperties": true,
                      "description": "поля правила из песочницы"
                    },
                    "trace": {
                      "type": "boolean"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "QUEUED",
                        "RUNNING",
                        "DONE",
                        "FAILED"
                      ]
                    },
                    "total": {
                      "type": "integer"
                    },
                    "done": {
                      "type": "integer"
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "created_by": {
                      "type": "string",
                      "nullable": true
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "started_at": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    },
                    "finished_at": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    },
                    "ahead": {
                      "type": "integer",
                      "description": "сколько прогонов в очереди перед этим: воркер берёт их по одному"
                    },
                    "results": {
                      "type": "array",
                      "description": "по объекту: QUEUED — ждёт очереди, DONE — итог прогона, FAILED — прогон на объекте не удался",
                      "items": {
                        "type": "object",
                        "required": [
                          "process_id",
                          "status"
                        ],
                        "properties": {
                          "process_id": {
                            "type": "string"
                          },
                          "object_id": {
                            "type": "string",
                            "nullable": true
                          },
                          "object_name": {
                            "type": "string",
                            "nullable": true
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "QUEUED",
                              "DONE",
                              "FAILED"
                            ]
                          },
                          "error": {
                            "type": "string",
                            "nullable": true
                          },
                          "elapsed_s": {
                            "type": "number",
                            "nullable": true
                          },
                          "result": {
                            "type": "object",
                            "nullable": true,
                            "additionalProperties": true,
                            "description": "base — записи рабочего правила, variant — варианта и разница, trace — кандидаты по стадиям"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
...
}